pyproject.toml
[virtualenvs]
create = true
in-project = true

[tool.poetry]
name = "20230903"
version = "0.1.0"
description = ""
authors = ["dearsecret <dearsecret1217@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
django-environ = "^0.10.0"
markdown = "^3.4.4"
django-filter = "^23.2"
pyjwt = "^2.8.0"
djangorestframework = "^3.14.0"
requests = "^2.31.0"
firebase-admin = "^6.2.0"
django-cors-headers = "^4.2.0"
pillow = "^10.0.1"
black = "^23.11.0"
pylance = "^0.8.17"
graphene-django = "^3.1.5"




[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
